apis/common/requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
boto3==1.28.61
setuptools==69.2.0
datasets==2.18.0
ragas==0.1.5
